requirements.txt
requests==2.31.0
openai==1.3.7
beautifulsoup4==4.12.2
jieba==0.42.1
numpy==1.24.3
typing-extensions==4.7.1
astroid==3.0.1
pylint==3.0.2
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0 